Reorganize files into directories
The updated workflow is

```
$ go build -o terraform-provider-slack
$ cd example
$ terraform init -plugin-dir ..
$ terraform plan
```
