+++ example/main.tf
@@ -1,3 +1,3 @@
-resource "slack" "slack" {
+resource "slack" "test" {
     hook_url = "https://api.slack.com"
 }
