+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"github.com/caulagi/terraform-provider-slack/slack"
 	"github.com/hashicorp/terraform-plugin-sdk/plugin"
 	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 )
@@ -8,7 +9,7 @@ import (
 func main() {
 	plugin.Serve(&plugin.ServeOpts{
 		ProviderFunc: func() terraform.ResourceProvider {
-			return Provider()
+			return slack.Provider()
 		},
 	})
 }
